Feat: Edit, Delete game group and fix game category issues

+++ backend/src/services/game-category.service.ts
@@ -1,8 +1,7 @@
-import { HTTP_STATUS } from "~/constants/httpStatus";
-import { ErrorWithStatus } from "~/models/Error";
 import gameCategoryRepository from "~/repositories/game-category.repository";
 import { CreateGameCategoryRequestBody, EditGameCategoryRequestBody } from "~/models/requests/game-category.request";
-import Helpers from "~/utils/helpers";
+import { HTTP_STATUS } from "~/constants/httpStatus";
+import { ErrorWithStatus } from "~/models/Error";
 
 class GameCategoryService {
     public create = async (data: CreateGameCategoryRequestBody) => {
@@ -12,13 +11,34 @@ class GameCategoryService {
     };
 
     public edit = async (id: string, data: EditGameCategoryRequestBody) => {
-        const { name, thumbnail, active } = data;
-        const slug = Helpers.generateSlug(name);
+        const category = await gameCategoryRepository.findById(id);
+        if (!category) {
+            throw new ErrorWithStatus({
+                status: HTTP_STATUS.NOT_FOUND,
+                message: "Danh mục không tồn tại!",
+            });
+        }
 
-        return await gameCategoryRepository.edit(id, { name, thumbnail, active, slug });
+        return await gameCategoryRepository.edit(id, data);
     };
 
     public delete = async (id: string) => {
+        const category = await gameCategoryRepository.findById(id);
+        if (!category) {
+            throw new ErrorWithStatus({
+                status: HTTP_STATUS.NOT_FOUND,
+                message: "Danh mục không tồn tại!",
+            });
+        }
+
+        const hasGameGroups = await gameCategoryRepository.hasGameGroups(id);
+        if (hasGameGroups) {
+            throw new ErrorWithStatus({
+                status: HTTP_STATUS.CONFLICT,
+                message: "Không thể xóa danh mục đang có group!",
+            });
+        }
+
         return await gameCategoryRepository.delete(id);
     };
 }
